fix(inputs.ping): Use deterministic packet IDs to prevent collisions
